Update dependencies
Can't boot due to Exposed issues: https://github.com/JetBrains/Exposed/pull/906

+++ loritta-discord/src/main/java/com/mrpowergamerbr/loritta/commands/vanilla/administration/WarnCommand.kt
@@ -112,9 +112,11 @@ class WarnCommand : AbstractCommand("warn", listOf("aviso"), CommandCategory.ADM
 					}
 				}
 
-				val warnCount = transaction(Databases.loritta) {
-					Warns.select { (Warns.guildId eq context.guild.idLong) and (Warns.userId eq user.idLong) }.count()
-				} + 1
+				val warnCount = (
+						transaction(Databases.loritta) {
+							Warns.select { (Warns.guildId eq context.guild.idLong) and (Warns.userId eq user.idLong) }.count()
+						} + 1
+						).toInt()
 
 				val punishments = punishmentActions.filter { it.warnCount == warnCount }
